refactor(map): move *.Types.Map* declarations to *.Types.Maps.Map*
