+++ Tests/YDotNet.Tests.Unit/Maps/ObserveDeepTests.cs
@@ -2,8 +2,8 @@ using NUnit.Framework;
 using YDotNet.Document;
 using YDotNet.Document.Cells;
 using YDotNet.Document.Events;
-using YDotNet.Document.Types;
 using YDotNet.Document.Types.Events;
+using YDotNet.Document.Types.Maps;
 
 namespace YDotNet.Tests.Unit.Maps;
