+++ Tests/YDotNet.Tests.Unit/Maps/ObserveTests.cs
@@ -1,8 +1,8 @@
 using NUnit.Framework;
 using YDotNet.Document;
 using YDotNet.Document.Cells;
-using YDotNet.Document.Types;
 using YDotNet.Document.Types.Events;
+using YDotNet.Document.Types.Maps;
 
 namespace YDotNet.Tests.Unit.Maps;
