+++ Tests/YDotNet.Tests.Unit/Maps/RemoveAllTests.cs
@@ -2,7 +2,7 @@ using NUnit.Framework;
 using YDotNet.Document;
 using YDotNet.Document.Cells;
 using YDotNet.Document.Transactions;
-using YDotNet.Document.Types;
+using YDotNet.Document.Types.Maps;
 
 namespace YDotNet.Tests.Unit.Maps;
